% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_all_cause_mortality.R
\docType{data}
\name{all_cause_mortality}
\alias{all_cause_mortality}
\title{Life table for the US in 2009}
\format{A data frame with 101 rows and 4 variables:
\describe{
\item{Age}{age, in years}
\item{Total}{mortality rate for the total population}
\item{Male}{mortality rate for males}
\item{Female}{mortality rate for females}
}}
\usage{
all_cause_mortality
}
\description{
A dataset with all-cause mortality rates from US 2009 life table estimates
}
\keyword{datasets}
