% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshr_from_ohr.R
\name{calc_avg_mort}
\alias{calc_avg_mort}
\title{Average background mortality}
\usage{
calc_avg_mort(n_age_init = 50, trial_time = 5, mu0 = NULL,
  alpha = NULL, hazard = c("empirical", "exponential", "linear",
  "geometric"), mortality = NULL)
}
\arguments{
\item{n_age_init}{initial age of the cohort in the trial}

\item{trial_time}{length of the trial}

\item{mu0}{age-independent mortality rate coefficient}

\item{alpha}{age-dependent mortality rate coefficient}

\item{hazard}{The hazard function to be used for background mortality.
Either empirical (e.g., from life tables) or a continous-time functional
form.}

\item{mortality}{A \code{data.frame} with two columns. The first column
should have age in yearly increments and the second column should have
the mortality rate corresponding to the age. If \code{NULL}, the function
will use \code{\link{all_cause_mortality}}.}
}
\value{
The average background mortality
}
\description{
\code{calc_avg_mort} calculates the average background mortality for a
cohort with initial age \code{n_age_init} in a trial with duration of
\code{trial_time} years for either empirical hazard (e.g.. from life tables)
or some continuous-time hazard functional forms.
}
