% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshr_from_ohr.R
\name{calc_dshr_from_ohr}
\alias{calc_dshr_from_ohr}
\title{Disease-specific hazard ration (dsHR) from an overall hazard ratio (oHR)}
\usage{
calc_dshr_from_ohr(ohr = 0.55, mu_Dis = 0.05, n_age_init = 50,
  trial_time = 5, mu0 = NULL, alpha = NULL, hazard = c("empirical",
  "exponential", "linear", "geometric"), mortality = NULL)
}
\arguments{
\item{ohr}{Overall hazard ratio}

\item{mu_Dis}{Disease-specific mortality rate}

\item{n_age_init}{Initial age of the cohort in the trial}

\item{trial_time}{Length of the trial}

\item{mu0}{Age-independent mortality rate coefficient}

\item{alpha}{Age-dependent mortality rate coefficient}

\item{hazard}{The hazard function to be used for background mortality.
Either empirical (e.g., from life tables) or a continous-time functional
form.}

\item{mortality}{A \code{data.frame} with two columns. The first column
should have age in yearly increments and the second column should have
the mortality rate corresponding to the age. If \code{NULL}, the function
will use \code{\link{all_cause_mortality}}.}
}
\value{
The disease-specific hazard ratio (dsHR)
}
\description{
\code{calc_dshr_from_ohr} calculates a dsHR from an oHR for a cohort with
initial age \code{n_age_init} in a trial with duration of \code{trial_time} years
with a disease-specific mortality rate, \code{mu_Dis}, under a specific
continuous-time hazard function for background mortality.
}
\section{Details}{

This function requires the average background mortality of the cohort, which
is computed with the function \code{\link{calc_avg_mort}}.
}

