% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshr_from_ohr.R
\name{calc_ohr_from_dshr}
\alias{calc_ohr_from_dshr}
\title{Overall hazard ratio (oHR) from a disease-specific hazard ratio (dsHR)}
\usage{
calc_ohr_from_dshr(dshr = 0.5, mu_Dis = 0.05, n_age_init = 50,
  trial_time = 5, mu0 = NULL, alpha = NULL, hazard = c("empirical",
  "exponential", "linear", "geometric"), mortality = NULL)
}
\arguments{
\item{dshr}{Disease-specific hazard ratio}

\item{mu_Dis}{Disease-specific mortality rate}

\item{n_age_init}{Initial age of the cohort in the trial}

\item{trial_time}{Length of the trial}

\item{mu0}{Age-independent mortality rate coefficient}

\item{alpha}{Age-dependent mortality rate coefficient}

\item{hazard}{The hazard function to be used for background mortality.
Either empirical (e.g., from life tables) or a continuous-time functional
form.}

\item{mortality}{A \code{data.frame} with two columns. The first column
should have age in yearly increments and the second column should have
the mortality rate corresponding to the age. If \code{NULL}, the function
will use \code{\link{all_cause_mortality}}.}
}
\value{
The overall hazard ratio (oHR)
}
\description{
\code{calc_ohr_from_dshr} calculates an oHR from a dsHR for a cohort with
initial age \code{n_age_init} in a trial with duration of \code{trial_time} years
with a disease-specific mortality rate, \code{mu_Dis}, under a specific
continuous-time hazard function for background mortality.
}
\section{Details}{

This function requires the average background mortality of the cohort, which
is computed with the function \code{\link{calc_avg_mort}}.
}

\examples{
n_age_init <- 50   # Initial age of cohort in trial
trial_time <- 5    # Length of clinial trial in years
ohr        <- 0.55 # Overall hazard ratio (oHR)
mu_Dis     <- 0.05 # Disease-specific mortality rate
\dontrun{
   dshr_emp <- calc_dshr_from_ohr(ohr = ohr, mu_Dis = mu_Dis,
   n_age_init = n_age_init, 
   trial_time = trial_time, 
   hazard = "empirical")
   dshr_emp
}
}
