% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticities.R
\name{comp_elast_gompz}
\alias{comp_elast_gompz}
\title{Elasticities of disease-specific hazard ratio (dsHR)}
\usage{
comp_elast_gompz(ohr = 0.55, mu_Dis = 0.05, n_age_init = 50,
  trial_time = 5, mu0 = NULL, alpha = NULL,
  hazard = c("exponential"), mortality = NULL)
}
\arguments{
\item{ohr}{Overall hazard ratio}

\item{mu_Dis}{Disease-specific mortality rate}

\item{n_age_init}{Initial age of the cohort in the trial}

\item{trial_time}{Length of the trial}

\item{mu0}{Age-independent mortality rate coefficient}

\item{alpha}{Age-dependent mortality rate coefficient}

\item{hazard}{The hazard function to be used for background mortality.
Either empirical (e.g., from life tables) or a continuous-time functional
form.}

\item{mortality}{A \code{data.frame} with two columns. The first column
should have age in yearly increments and the second column should have
the mortality rate corresponding to the age. If \code{NULL}, the function
will use \code{\link{all_cause_mortality}}.}
}
\value{
Elasticities of dsHR with respect to trial parameters.
}
\description{
\code{comp_elast_gompz} computes the elasticities of dsHR with respect to
parameters obtained from a clinical trial
}
