% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshr.R
\docType{package}
\name{dshr}
\alias{dshr}
\alias{dshr-package}
\title{\code{dshr} package}
\description{
Disease-Specific Hazard Ratio from Overall Hazard Ratio
}
\details{
See the README on
\href{https://github.com/feralaes/dshr#readme}{GitHub}
}
