% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshr_from_ohr.R
\name{est_hazard_params}
\alias{est_hazard_params}
\title{Estimation of hazard functional forms for background mortality}
\usage{
est_hazard_params(n_age_init = 50, trial_time = 5,
  hazard = c("exponential", "linear", "geometric"), mortality = NULL)
}
\arguments{
\item{n_age_init}{Initial age of the cohort}

\item{trial_time}{Duration of clinical trial in years}

\item{hazard}{The hazard function to be used for background mortality.
Either empirical (e.g., from life tables) or a continuous-time functional
form.}

\item{mortality}{A \code{data.frame} with two columns. The first column
should have age in yearly increments and the second column should have
the mortality rate corresponding to the age. If \code{NULL}, the function
will use \code{\link{all_cause_mortality}}.}
}
\value{
A vector with an age-independent mortality rate coefficient, \code{mu0} and
an age-dependent mortality rate coefficient, \code{alpha}.
}
\description{
\code{est_hazard_params} estimates the parameters of different
continuous-time hazard functional forms for background  mortality for a
cohort with initial age \code{n_age_init} in a trial with duration of
\code{trial_time} years.
}
\examples{
n_age_init <- 50   # Initial age of cohort in trial
trial_time <- 5    # Length of clinial trial in years
\dontrun{
   v_coef_hazard <- est_hazard_params(n_age_init = n_age_init, 
   trial_time = trial_time, 
   hazard = "exponential")
   v_coef_hazard
}
}
